+++ R/tm_t_events_patyear.R
@@ -406,10 +406,12 @@ srv_events_patyear <- function(input,
     validate(
       need(is_character_single(input_aval_var), "`Analysis Variable` should be a single column."),
       need(is_character_single(input_events_var), "Events variable should be a single column."),
-      need(input$conf_method, "`CI Method` field is not selected"),
+      need(input$conf_method, "`CI Method` field is not selected."),
       need(input$time_unit_output, "`Time Unit for AE Rate (in Patient-Years)` field is empty."),
       need(input[[extract_input("paramcd", paramcd$filter[[1]]$dataname, filter = TRUE)]],
-        "`Select an Event Type Parameter is not selected.")
+        "`Select an Event Type Parameter is not selected."),
+      need(!any(is.na(anl_m$data()[[input_events_var]])),
+        "`Event Variable` for selected parameter includes NA values.")
     )
 
     NULL